refactoring and added store 3d scans to fgs plan
