+++ src/artemis/fast_grid_scan_plan.py
@@ -1,6 +1,7 @@
 import os
 import sys
 from collections import namedtuple
+from selectors import EpollSelector
 
 sys.path.append(os.path.join(os.path.dirname(__file__), "..", ".."))
 
@@ -15,7 +16,7 @@ from ophyd.log import config_ophyd_logging
 from src.artemis.devices.eiger import EigerDetector
 from src.artemis.devices.fast_grid_scan import FastGridScan, set_fast_grid_scan_params
 from src.artemis.devices.zebra import Zebra
-from src.artemis.ispyb.store_in_ispyb import StoreInIspyb
+from src.artemis.ispyb.store_in_ispyb import StoreInIspyb2D, StoreInIspyb3D
 from src.artemis.nexus_writing.write_nexus import NexusWriter
 from src.artemis.parameters import SIM_BEAMLINE, FullParameters
 from src.artemis.zocalo_interaction import run_end, run_start
@@ -35,9 +36,16 @@ config_ophyd_logging(file="/tmp/ophyd.log", level="DEBUG")
 def run_gridscan(
     fgs: FastGridScan, zebra: Zebra, eiger: EigerDetector, parameters: FullParameters
 ):
-    ispyb = StoreInIspyb("config", parameters)
-    _, datacollection_id, datacollection_group_id = ispyb.store_grid_scan()
-    run_start(datacollection_id)
+    config = "config"
+    if fgs.is_3d_grid_scan():
+        ispyb = StoreInIspyb3D(config)
+    else:
+        ispyb = StoreInIspyb2D(config)
+    datacollection_ids, _, datacollection_group_id = ispyb.store_grid_scan(parameters)
+
+    for id in datacollection_ids:
+        run_start(id)
+
     # TODO: Check topup gate
     yield from set_fast_grid_scan_params(fgs, parameters.grid_scan_params)
 
@@ -50,13 +58,16 @@ def run_gridscan(
         yield from do_fgs()
 
     current_time = ispyb.get_current_time_string()
-    ispyb.update_grid_scan_with_end_time_and_status(
-        current_time,
-        "DataCollection Successful",
-        datacollection_id,
-        datacollection_group_id,
-    )
-    run_end(datacollection_id)
+    for id in datacollection_ids:
+        ispyb.update_grid_scan_with_end_time_and_status(
+            current_time,
+            "DataCollection Successful",
+            id,
+            datacollection_group_id,
+        )
+
+    for id in datacollection_ids:
+        run_end(id)
 
 
 def get_plan(parameters: FullParameters):
